Split composer data models
