+++ assets/js/publish/transports/connect-transport.js
@@ -1,4 +1,4 @@
-import { CONNECT_PUBLISH_MESSAGE_TYPE } from '../settings-store.js?v=press-system-v3.4.21';
+import { CONNECT_PUBLISH_MESSAGE_TYPE } from '../settings-store.js?v=press-system-v3.4.22';
 
 export function serializeConnectPublishFile(file) {
   const out = {
