+++ assets/js/theme-manager.js
@@ -1,5 +1,5 @@
-import { t } from './i18n.js?v=press-system-v3.4.21';
-import { loadPressSystemManifest, satisfiesSemverRange } from './press-version.js?v=press-system-v3.4.21';
+import { t } from './i18n.js?v=press-system-v3.4.22';
+import { loadPressSystemManifest, satisfiesSemverRange } from './press-version.js?v=press-system-v3.4.22';
 import { unzipSync, strFromU8 } from './vendor/fflate.browser.js';
 
 const THEME_ROOT = 'assets/themes';
